Make references proper object references
We had the option to either leave references as soft, string only references or
go all out and directly reference the actual object of the other element. The
latter makes for a much nicer API to work with, because you can easily travel
through the data if you wanted to. The former obviously means less work on the
library side. I think the trade-off is worth it to have a nicer API but also
takes a hit in the initial load performance because of all this shifting back
and forth. Current algorithm takes O(2n) which isn't completely horrible, but
isn't great either, though I doubt I can get this down to a single pass.

This also introduces some better representations for data, such as the status of
 a task/project and the relative duration for things like when the next
repetition is. However, there are some uncertainties here as I couldn't quite
figure out what ever possible value means yet.