Be more helpful when creating a storage from a path
It will now error out if the path specified cannot be used to find the
omnifocus database. If the path does not end in ".ofocus" it will assume that
the parent directory was presented and will try to construct the correct path
itself. When the resulting path still doesn't lead to a correct directory, it
will provide a proper exception to help the user of the library figure out what
 is wrong.