Provide functionality to create own omni device
To create changes, they must be attached to a device - the author of those
changes. Of course, one could use one of the already registered devices, but
that could lead to issues as the author would not even know of the changes! To
avoid this, we allow for creating custom devices, that may later be used to
create changes.

The device requires us to fill out a lot of information of the device, most of
which we can easily supply. However, to make sure we get as close to the real
data and avoid confusing omnifocus, we use OSHI to provide some of the missing
hardware information (specifically: CPU information). If this proves to be
irrelevant, we might chose to remove it again.