Implement basic CLI functionality
This contains a lot of changes that just accumulated over time. One of them
being a proper cli interface that allows for sub commands, for example
`focus storage list-changesets`. For this, we switched to clikt. To allow for
nicer cli usage, this also added configuration support. There, the location and
password of the storage can be saved so that they will be used instead of having
 to pass them as an argument.

This also creates the foundation for creating own changesets as the generation
of them has been done.