Use a simpler configuration library
This switches the configuration from `konf` to `toml4j`. While the latter isn't
specifically a configuration library, we use it as that. The reason for this
change is simply put: performance. Currently with konf, doing a query takes ~3
seconds on an encrypted store. A second is used for generating the decryption
information, another is for doing command parsing, creating the view, etc. The
third second is used _just_ for parsing & loading the configuration, which _way_
 too much for this use case. While it would've been nice to be able to use
environment vars to override the configuration, it's really not needed. We
already provide command line options to cover this and thus don't really provide
 and advantage.

By using a simple toml configuration library, we basically cut this down to
1/10th of the time, making the total time a command takes to execute to be
around 2seconds. Which is still not great, but there doesn't seem to be much
that can be done here; we need to re-create the aes key on every startup since
we wouldn't want to save that somewhere.