Make operation of storage elements a property
This moves the operation property from the constructor to a normal
property. The reason for this change is to allow for comparisons to
work without comparing the operation, as that is not helpful. We could
create a separate comparison function, but that would be a lot more
work than necessary.