Handle reference elements properly in changesets
Currently we just ignore reference elements when going though the
changesets; we assume that everything is correct. This now changes that
by actually checking if current state we have matches the state in the
reference. However, this makes no attempt to completely handle the
reference element, such as when creating a new changeset. It is not
quite clear to me _when_ the reference is used and when it isn't.

Reference elements - for some reason - use an inner container that contains the
content of the current element values. Therefore when reading/writing
references, we need to consider this encapsulation.