Add configuration option to refresh configured device
OmniFocus saves the last seen ID for every device. If all devices have seen a
certain change, they may device to compress all changes up to and including that
 change into the initial changeset. That eliminates the small changes that
happened over time and would no longer be necessary (i.e. name changes for
things). But, as said, this only works if all devices have seen the change,
which we currently do not update on queries!

We don't really care about this; we currently re-read the whole store each time
we do any action. As a result, we'd just update the id of the configured device
to the most recent one.