Differentiate between changesets and their descriptors
This refactors the code to make a clear distinction between the "descriptors"
and the actual changesets. Since loading changesets is comparatively expensive,
there are many use cases where this is not really necessary and would thus just
waste time. As a result, we provide a descriptor for changesets that then can be
 used to get the actual contents of the changeset while itself being fast to
load and staying small. The descriptor contains all the information we have
about a changeset before parsing: timestamp, id & previous id. Already this
allows narrowing down the possible changesets we could want (e.g. all
changesets newer than X) as well as allow us to properly load a changeset for a
given id - by finding the descriptor and then loading the changeset via that.

Previously there already was a bit of abstraction for this part but it was not
very well thought through and was a bit confusing and/or intertwined with the
storage implementation. This lays the foundation to read updated changes from
the store.