Only use saved password if path actually matches
Currently we automatically use the saved password even if the user manually
specified a storage location in the command. However, that may be a different
storage and thus uses a different password. To prevent that, we also have to
ensure the paths are the same before continuing with the password.