Add functionality to watch for changes to the store
This allows a user to register a changeset watcher on the store to get notified
about changes, such as a new changeset or old ones being removed. However, this
is a very basic initial implementation which may not handle certain cases yet,
specifically including when some get merged into the base changeset.

Additionally, as noted in the implementation comment in
`WatchserviceChangesetWatcher` it does not work with sync services that use
temporary files (like in the format of `._xyz`) and then rename them as there's
no java API that covers renames. In such a case, one would need to use the
`PollingChangesetWatcher` instead.